Upgrading «scodec» to version 1.5.0
This allows to take advantage of the scodec.Err error type returned by
codecs instead of the basic String of previous versions. Currently I have
just used the Err.apply to fulfill this upgrade. But I'll have to derive
my own errors from trait Err for it to be usefull.