'TCPTransport' Refactoring
The 'TCPTransport' abstract class doesn't take the client actor as a
construction parameter anymore. The client is now supposed to be the
actor creating an instance of a 'TCPTransport' derived class. For the
specifications (Specs2) to work, some boilerplate is required to provide
a fake parent to this instance. see https://gist.github.com/archie/7952671
for more examples on that.