Revisiting how publication works
I changed my mind and considered, for the time being, to make publication
handling far simpler in the code. The message identifier for QoS above 0
is now required. Every publication acknowlegement comes with the original
message identifier in it. So there is no more message identifier generator
within 'Protocol'.

Moreover publication retries are the responsibility of the client code, no
logic is build into 'Protocol' for that purpose. This makes the API very
low level! Maybe I'll revisit that aspect later but for now I want it to
be kept simple and just work...