The message identifier is not always here!
The 'Publish' frame has a variable header where the presence of the
message identifier depends on the QoS provided by the header. I thought
this dependence, according to the MQTT specification was valid for other
frames, but this is not the case. Just because, for instance, some responses
ignore to set the QoS!