Complete rewrite the MQTT keep alive
The purpose of this rewrite is to avoid sending PINGREG messages to the
broker when there is no real need. The previous implementation used two
timers. One periodically triggering a 'SendKeepAlive' message every
'keepAlive' seconds, the other signaling the PINGRESP timeout after
'keepAlive' seconds.

Obviously there was no need for two timers with the same delay. Moreover
the first timer triggered a PINGREQ message even if traffic hapened from
the client to the broker.

Now there is only one timer with a smarter reschedule policy taking into
account the timestamp of the last message sent, the current time and the
keep alive value. All the logic has been transfered to the 'Protocol'
trait still trying to keep side effects in the 'Transport' class.