Lifting 'MessageIdentifier' into the API
Finally, 'MessageIdentifier' is renamed 'MQTTMessageId' and lifted into
the API where it seems to belong. This is a 'contract' class not a low
level thing related to protocol frames. Its 'raison d'être' is only to
enforce some rules and guide the end user. The same will soon happen with
the 'CaseEnum' derived classes!