Using the 'Scalaz' state monad
What was previously called 'State' has become 'Registers' now. The idea
is to corner the impure effects in the Akka message handlers as much as
possible. Well... The truth is that I wanted to tinker with the state
monad, even if this was not really necesary. Just to learn about its use!